requirements.txt
# Core Flask Dependencies
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
Werkzeug==3.0.3
Flask-SocketIO==5.3.6
python-socketio==5.11.3

# MongoDB Dependencies
Flask-PyMongo==2.3.0
pymongo==4.8.0
dnspython==2.6.1

# File Upload Dependencies
Flask-Uploads==0.2.1
Flask-Reuploaded==1.4.0
pillow==10.4.0

# Security Dependencies
bcrypt==4.2.0

# Environment and Configuration
python-dotenv==1.0.1

# Template and Form Dependencies
Jinja2==3.1.4
WTForms==3.1.2
MarkupSafe==2.1.5

# Utility Dependencies
click==8.1.7
blinker==1.8.2
colorama==0.4.6
itsdangerous==2.2.0